requirements.txt
arrow                 
asgiref               
binaryornot           
certifi
chardet
charset-normalizer
click
colorama
cookiecutter
Django
djangorestframework
django-browser-reload 
django-debug-toolbar
django-filter
django-redis
django-tailwind
idna
markdown
markdown-it-py        
MarkupSafe            
mdurl                 
pillow                
pip                   
psycopg2              
Pygments              
python-dateutil       
python-environ        
python-slugify        
PyYAML                
requests              
rich                  
six                   
sqlparse              
text-unidecode        
types-python-dateutil 
tzdata                
urllib3               